DEV: Update assistant component
