+++ app/assets/javascripts/discourse/app/components/search-menu/assistant-item.js
@@ -1,70 +1,76 @@
-createWidget("search-menu-assistant-item", {
-  tagName: "li.search-menu-assistant-item",
+import Component from "@glimmer/component";
+import { inject as service } from "@ember/service";
+import { action } from "@ember/object";
+import { bind } from "discourse-common/utils/decorators";
+import { tracked } from "@glimmer/tracking";
 
-  html(attrs) {
-    const prefix = attrs.prefix?.trim();
-    const attributes = {};
-    attributes.href = "#";
+export default class AssistantItem extends Component {
+  constructor() {
+    super(...arguments);
 
-    let content = [
-      h(
-        "span",
-        { attributes: { "aria-label": I18n.t("search.title") } },
-        iconNode(attrs.icon || "search")
-      ),
-    ];
+    //const prefix = attrs.prefix?.trim();
+    //const attributes = {};
+    //attributes.href = "#";
 
-    if (prefix) {
-      content.push(h("span.search-item-prefix", `${prefix} `));
-    }
+    //let content = [
+    //h(
+    //"span",
+    //{ attributes: { "aria-label": I18n.t("search.title") } },
+    //iconNode(attrs.icon || "search")
+    //),
+    //];
 
-    if (attrs.withInLabel) {
-      content.push(h("span.label-suffix", `${I18n.t("search.in")} `));
-    }
+    //if (prefix) {
+    //content.push(h("span.search-item-prefix", `${prefix} `));
+    //}
 
-    if (attrs.category) {
-      attributes.href = attrs.category.url;
+    //if (attrs.withInLabel) {
+    //content.push(h("span.label-suffix", `${I18n.t("search.in")} `));
+    //}
 
-      content.push(
-        this.attach("category-link", {
-          category: attrs.category,
-          allowUncategorized: true,
-          recursive: true,
-          link: false,
-        })
-      );
+    //if (attrs.category) {
+    //attributes.href = attrs.category.url;
 
-      // category and tag combination
-      if (attrs.tag && attrs.isIntersection) {
-        attributes.href = getURL(`/tag/${attrs.tag}`);
-        content.push(h("span.search-item-tag", [iconNode("tag"), attrs.tag]));
-      }
-    } else if (attrs.tag) {
-      if (attrs.isIntersection && attrs.additionalTags?.length) {
-        const tags = [attrs.tag, ...attrs.additionalTags];
-        content.push(h("span.search-item-tag", `tags:${tags.join("+")}`));
-      } else {
-        attributes.href = getURL(`/tag/${attrs.tag}`);
-        content.push(h("span.search-item-tag", [iconNode("tag"), attrs.tag]));
-      }
-    } else if (attrs.user) {
-      const userResult = [
-        avatarImg("small", {
-          template: attrs.user.avatar_template,
-          username: attrs.user.username,
-        }),
-        h("span.username", formatUsername(attrs.user.username)),
-        attrs.suffix,
-      ];
-      content.push(h("span.search-item-user", userResult));
-    } else {
-      content.push(h("span.search-item-slug", attrs.label || attrs.slug));
-    }
-    if (attrs.extraHint) {
-      content.push(h("span.extra-hint", attrs.extraHint));
-    }
-    return h("a.widget-link.search-link", { attributes }, content);
-  },
+    //content.push(
+    //this.attach("category-link", {
+    //category: attrs.category,
+    //allowUncategorized: true,
+    //recursive: true,
+    //link: false,
+    //})
+    //);
+
+    //category and tag combination
+    //if (attrs.tag && attrs.isIntersection) {
+    //attributes.href = getURL(`/tag/${attrs.tag}`);
+    //content.push(h("span.search-item-tag", [iconNode("tag"), attrs.tag]));
+    //}
+    //} else if (attrs.tag) {
+    //if (attrs.isIntersection && attrs.additionalTags?.length) {
+    //const tags = [attrs.tag, ...attrs.additionalTags];
+    //content.push(h("span.search-item-tag", `tags:${tags.join("+")}`));
+    //} else {
+    //attributes.href = getURL(`/tag/${attrs.tag}`);
+    //content.push(h("span.search-item-tag", [iconNode("tag"), attrs.tag]));
+    //}
+    //} else if (attrs.user) {
+    //const userResult = [
+    //avatarImg("small", {
+    //template: attrs.user.avatar_template,
+    //username: attrs.user.username,
+    //}),
+    //h("span.username", formatUsername(attrs.user.username)),
+    //attrs.suffix,
+    //];
+    //content.push(h("span.search-item-user", userResult));
+    //} else {
+    //content.push(h("span.search-item-slug", attrs.label || attrs.slug));
+    //}
+    //if (attrs.extraHint) {
+    //content.push(h("span.extra-hint", attrs.extraHint));
+    //}
+    //return h("a.widget-link.search-link", { attributes }, content);
+  }
 
   click(e) {
     const searchInput = document.querySelector("#search-term");
@@ -77,5 +83,5 @@ createWidget("search-menu-assistant-item", {
     });
     e.preventDefault();
     return false;
-  },
-});
+  }
+}
