+++ app/assets/javascripts/discourse/app/components/search-menu/results/assistant.hbs
@@ -1,0 +1,3 @@
+<ul class="search-menu-assistant">
+
+</ul>
